removed graphical list bullet
Unordered lists used a graphic for the list bullet. Using the same technique
as in the ordered lists we don't need an image just to have the color. This
has two advantages:

1. The list bullets still have the correct color when colors are adjusted
in style.ini

2. List bullets grow and shrink correctly with the fontsize

It has a disadvantage as well: All list items need to be marked up like this:

<li><div class

darcs-hash:20060121121435-7ad00-c6455f9b58c6e78b630c1121499e2052525fb556.gz