Add the Universal Wiki Edit Button to the deafault template
The "Universal Wiki Edit Button" is a movement to make Wikis easily
recognizable to users regardless of the used engine or template. The
aim is to create an icon which is as easy to recognize as the now well
known RSS icon. The hope is to spread the Wiki idea and make more people
aware of the use of Wikis outside the scope of Wikipedia.

The logo is added through a background CSS defined in UWEB.css

More info about the Universal Wiki Edit Button can be found at
http://www.aboutus.org/uweb - you may also add your thoughts about it
there.

darcs-hash:20071026213115-7ad00-6101fbc3449b4426f795c60775c874eb0e53bb82.gz