experimental Flash based multi upload
This patch adds experimental support for uploading multiple files in one go.
This is achieved by using Flash for selecting multiple images and intitiating
the HTTP upload.

When Flash 8 or higher is detected, an additional icon is added to the usual
upload form. Clicking it will swith the form to the Flash based upload queue.

Things that need work:

  * Better Icon
  * Feedback if Flash detection works on all Flash 8 supported platforms
  * Progress feedback seems not to work on Linux (might be Adobe bug)
  * No final feedback how many images were uploaded correctly

The flash sources are located in lib/_fla/

Any feedback and help would be appreciated.

darcs-hash:20080606222849-7ad00-738083445af275752aaebc29bfa51430f3d94459.gz