Link Wizard added
Ignore-this: c15561aa909f921f7845576378851b93

This adds a new link wizard to the toolbar which helps users to find the page the want to link to.

Pages can be found by a simple page name search or by browsing the
existing namespaces.

This is the first checkin. Some cleanup and MSIE compatibility checks
remain.

note: development was part of the ICKE 2.0 project see
      http://www.icke-projekt.de for info

darcs-hash:20090812102302-6e07b-fcc564fcaf2ed6aa832918870dd0f92607748687.gz